Create HomePage component.
Moved the home page to a separate file to keep it organized.
